requirements.txt
# SamvidhanAI - Constitutional RAG System Dependencies

# LangChain Core
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38

# Google Gemini Integration
langchain-google-genai==1.0.1
google-generativeai==0.4.1

# Vector Database
chromadb==0.4.24

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.11.0

# Web Framework
streamlit==1.32.2

# Utilities
python-dotenv==1.0.1
requests==2.31.0

# Data Processing
numpy==1.26.4
pandas==2.2.1

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
